pyproject.toml
[tool.poetry]
name = "net-shop"
version = "0.1.0"
description = ""
authors = ["m1khaeel <mikestepanoff44@gmail.com>"]
readme = "README.md"
packages = [{include = "net_shop"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
djoser = "^2.1.0"
pytest-django = "^4.5.2"
prospector = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
